Library: Fix application crash caused by SQL error
When the SQLite database does not exist and the user adds a component to
a schematic, an SQL error appears and the whole application crashes.
The thrown exception is caused by non-existing tables in the database.
The crash is caused by the method Library::getElementFilePathsFromDb()
which is mistakenly marked with the "noexcept" keyword although it can
thow exceptions.

This commit does:
- Avoid the thrown exception by creating all non-existing tables right
  after opening the database (ctor of Library).
- Fix the application crash by correcting the exception specifier of
  Library::getElementFilePathsFromDb().
- Make private constructors/operators public and add "=delete" specifier

Fixes #78