Travis: Suppress console output of git commit/push
We use an encrypted string in .travis.yml to get write access to our
doxygen output repository (https://github.com/LibrePCB/LibrePCB-Doxygen).
But the encryption is useless if "git push" prints the connection string
to the console output ;)

Because the old connection string is now public, I created a new one.