Allow to create projects without schematics/boards
When creating a new project, schematics and boards are no longer
created/added automatically. They must now be created and added to the
project manually. This makes it possible to create projects without any
schematics and/or boards, and it allows to manually specify the name of
the first schematic and/or board of a project.