Travis: Run unit tests with "xvfb-run"
As the unit tests now use the QtWidgets module, an X server is required
to run the tests. Otherwise we would get the following error message:
"QXcbConnection: Could not connect to display"

To get the unit tests running on Travis-CI, we will now just use "xvfb"
(X window virtual framebuffer).