Move defines from `common.pri` to `librepcbcommon.pro`
Our preprocessor defines in `common.pri` were accessible in all source
files of the whole project. So for example to read the application's
version number, one could read it from the `Application` class or
directly from the define `APP_VERSION`. Now these defines are moved to
the project file `librepcbcommon.pro`, so the scope of them is much
smaller and the design is cleaner.

Changes:
- Move preprocessor defines from `common.pri` to `librepcbcommon.pro`
- Replace defines `APP_VERSION_MAJOR/MINOR/PATCH` with `APP_VERSION`
- Add define `FILE_FORMAT_VERSION` and set it to `0.1`
- Show app version in about dialog (in addition to the Git version)
- Make static functions of class `Application` non-static
- Redefine macro `qApp` to make it working with our class `Application`
- Update unit test `ApplicationTest`