Move method FilePath::mkPath() to FileUtils::makePath()
The class FilePath should only provide methods to deal with file paths,
and not to make changes in the file system. For this purpose we have
the class FileUtils, so the method is now moved to there.

In addition, the method FileUtils::makePath() now throws an exception in
case of an error (instead of returning a bool value). This makes error
handling easier.