Editors: Open context menu even if cursor was moved
Until now, the context menu in the GraphicsView of the SchematicEditor
and BoardEditor did not appear if the cursor was moved while the button
was pressed. This was necessary because the right mouse button was also
used for panning.

As the right mouse button is no longer used for panning, this is not
necessary anymore, so I removed this (sometimes annoying) behavior.