Refactor runtime resources system
The filepath to runtime resource files is no longer configurable by
preprocessor defines because this was too error-prone and made it
difficult to create packages for distributing the application.

Runtime resources must now be located at ../share/librepcb relative to
the executable. This is much more portable.

In addition, the directories ./res/ and ./packaging/ were replaced by
the new directory ./share/, which contains all resource files in a
directory structure which can be simply copied to /usr/share/ to install
them on a Unix/Linux system.