Deploy Qt translation files with the new resources system
Qt translation files (*.qm) are no longer statically compiled into the
librepcb executable, but just copied to the runtime resources directory
./share/ instead. This simplifies the deployment of translation files.

In addition, the change in librepcb.pro speeds up the build process
if there are no changes in source files, so the linker no longer has to
be invoked (before this change, it was always invoked, even if there
were no changes in source files). This saves ~2s in build time.