Fix random order of attributes in XML files
QDomDocument was not able to produce deterministic XML output because it
stores attributes in a QHash and thus their order is random. This is
very bad for version control systems because the diffs on these XML
files is unnecessary large.

Now a QXmlStreamWriter is used for serialization, which dooes not have
this limitation and thus the generated XML files are deterministic.