ComponentInstance: Force all symbols to be on same schematic
Actually it would be possible to place the symbols of a component on
different schematics. But maybe some time this will no longer be
possible due to the concept of hierarchical sheets, sub-circuits or
something like that. To make the later project upgrade process (as
simple as) possible, we introduce this restriction already from now on.