Adjust representation of arc segments in paths
The angle of a vertex specified the angle from the last vertex to that
vertex until now. But this does not really make sense because the angle
of the first vertex in a path is always useless.

Now the angle of a vertex specifies the angle from that vertex to the
next vertex. For paths which are implicitly considered as closed (even
if last vertex != first vertex), the last vertex can now be omitted
because the angle is already contained in the vertex before.