Integrate ProjectLibraryUpdater into LibrePCB application
Instead of a standalone, very technical application for developers, the
Project Library Updater is now simplified and integrated into the
LibrePCB application. So project libraries can now be updated directly
from within the Control Panel, Schematic Editor or Board Editor.

But note that the update mechanism is still very stupid (it just copies
directories from the workspace to the project) and thus the usability
and functionality is very limited.