Stop copying "share" directory to build output
The copy action of the "share" directory had several drawbacks:

- Increased build time
- Unnecessary complex build process
- Workaround on macOS needed to fix wrong relative path to executable

Now the application determines at startup whether it's running from
inside the build output directory (i.e. it's a developer build) or from
outside (i.e. it's a packaged/installed build). If it's a developer
build, the "share" directory of the repository root is used. If not, the
directory "../share" relative to the executable is used.