Change build output path to "./output"
Previously the build output path was "./generated/<unix|mac|windows>/",
depending on the operating system. Now it is simplified to "./output",
(independent of the operating system) because it doesn't make sense to
use different output paths for different systems.

In addition, the variable "GENERATED_DIR" in all *.pro files is no
longer needed.