Use separate *.sqlite files for each database version
Until now, the workspace library database is stored in the file
"cache.sqlite" in the workspace. And the database itself contains a
version number which is intended to be incremented if we change the
database layout.

But if you use two different LibrePCB application versions on the same
workspace, this doesn't work well in every case. The newer application
would upgrade the database on the first application start, but the older
application may no longer be able to work with that database version. So
generally we must introduce database format changes only with new major
versions of the application, which decreases our flexibility in minor
releases.

This change puts the database version number into the filename of the
database, i.e. the file "cache.sqlite" will be renamed to
"cache_v1.sqlite" and every future database version will use a separate
file. This way the issue as described above is fixed, and we still have
the flexibility to introduce breaking database changes at any time, even
in minor releases.