Fix state saving to occur when window is closed
Presently, there's a bug where I close a schematic editor or board
editor, and the state is not saved. Later, the state is saved in the
destructor called with `delete` in `~ProjectEditor`.

Because the window is already closed, all the widgets are gone in the
saved state. When I reopen the window, I do not have any menus or dock
widgets.