Add pre-selected devices/packages to project library
When adding a component with a pre-selected device to a schematic,
currently the device and its package are not copied to the project's
library. Thus the pre-selected device would not be available for adding
it to a board if the workspace does not contain the device (anymore).

Now the pre-selected devices and their packages are immediately added
to the project's library, so it's guaranteed to have them available for
adding to a board later, even if the workspace does not contain them.

This is needed for example to show the name of the pre-selected devices
instead of only their UUID (currently only the UUID was stored in the
project, but the name had to be fetched from the workspace library).