TransactionalFileSystem: Use callback to ask for restoring
Instead of passing an enum for choosing the restore mode (yes/no/ask/
abort), a callback can now be passed. The callback then returns a bool
to choose whether the backup should be restored or not. This way the
message box has to be implemented in higher level classes, and thus
can be more specific.