Replace UUID validation RegEx with manual validation (#651)
When analyzing a callgrind profile of a full library rescan, I noticed
that there are a lot of CPU instructions coming from libpcre2, with the
call originating in the UUID validation code.

Previously, the regular expression used for UUID validation was parsed
from a string for every invocation of the validation function. This is
unnecessary. A library rescan will validate thousands of UUIDs (roughly
250k calls to `Uuid::isValid` with 200'000 instructions per call on my
laptop for a library rescan). By replacing the RegEx with a manually
loop-unrolled validation function, I achieved a 20-40% speedup for a
full library scan on my computer (which has a fast SSD). The CPU
instructions per call to `isValid` were reduced from 187'341 to 767.

For more details, please refer to the pull request discussion:
https://github.com/LibrePCB/LibrePCB/pull/651