SchematicPagesDock: Disable wrapping of schematic pages
If there is not enough space available to display the schematic pages
in a vertical layout, QListWidget was allowed to wrap pages to the
right, i.e. using a horizontal layout. In some situations (depending on
the width of the dock widget), the wrapped pages were not visible
anymore, so it looked like some schematic pages disappeared. By
disabling wrapping, QListWidget always uses a vertical layout.