SchematicEditor: Refactor whole state machine
Migrating to the simpler, cleaner FSM design pattern as already used in
the library editor. The reduced complexity should make it easier to
implement new schematic editor features. In addition, the naming of the
classes is now more expressive.

The behavior of the schematic editor should not be affected by this
change, it's only a code refactoring.