BoardEditor: Refactor whole state machine
Migrating to the simpler, cleaner FSM design pattern as already used in
the library editor and schematic editor. The reduced complexity should
make it easier to implement new board editor features. In addition, the
naming of the classes is now more expressive.

The behavior of the board editor should not be affected by this change,
except a few tiny fixes.