Schematic editor: Fix leaving tool with right click + move
When pressing the right mouse button while moving the cursor, each
schematic editor tool unintentionally exited because the event was not
handled by the tool. So the event was handled by the state machine,
which switches to the last active tool by the right mouse button click.

Now the event is accepted by the corresponding tools to avoid this
issue.