Fix possibly outdated displayed net names in boards
The displayed net names within vias and pads, and the net names within
the hints of vias, pads and traces were not updated properly if a net
signal was renamed. So the displayed net names in boards was wrong. Now
the event of renaming a net signal is forwarded to vias, pads and traces
to enforce updating the displayed texts and hints accordingly.