LibraryDownload: Remove ZIP before starting download
If the temporary ZIP file already exists before starting to download a
library, the download failed with a "file already exists" error. This
can happen for example when closing LibrePCB while a library download
was in progress. Afterwards, it was not possible anymore to download
the library with the library manager.

Now the ZIP file will be removed before starting the download to fix
this issue.