Board editor: Fix staying in the "place device" tool
After a device has been placed on the board, the "place device" tool
was not left automatically. So the user had to manually click on the
"select" tool button to be able to select and move items. Now the
"select" tool is automatically entered after placing a device.

The bug was probably introduced by the last board editor FSM
refactoring.