Improve error messages when project library update fails
If the pinout of a device or component was changed after adding such
elements to a project, the project library updater might fail due to
invalid connections when running the next time. This is expected
behavior (thus a warning appears in the library editor), but now the
user at least gets a much more expressive error message. This way it's
easier to determine which library element caused the error.

Since these are still pretty low level error messages, I didn't made
them translatable, as decided in https://librepcb.discourse.group/t/119.