Unit Test: Fix failing testGetProcessNameByPid
On a windows 7-64 computer with Qt 5.15.0 the unit test of getProcessNameByPid failed with a
...C++ exception with description "QueryFullProcessImageNameW() failed with error 31." thrown in the test body...
error 31 is defined as ERROR_GEN_FAILURE (A device attached to the system is not functioning) which is not
really pointing to a solution.

The exception was thrown because the process handle returned by OpenProcess(..., pid) was not nullptr but
GetExitCodeProcess(pid) would have returned 62097, a magic value used by Qt when kill() is called and this was
indicating the process is killed and will disappear soon.
Not actually fixed the unit test code but the getProcessNameByPid code for checking this special end-of-process-life
condition and when detected return an empty string as the process name.