Allow creating unconnected board net segments (traces/vias)
This allows to create traces and vias which are not assigned to any net.
The main purpose is to provide the ability to create plated through
holes for mechanical purposes only. A positive side effect is that the
"add via" tool in the board editor is no longer rejected if the
schematic is empty (i.e. no nets available) - now this tool works as
usual since it is able to create unconnected vias.