SQLiteDatabaseTest: Simplify flaky concurrency test case
After several years of annoying issues with this unit test on Windows
(because Windows cannot reliably remove files, run multiple threads at
the same time, and because Windows is incredibly slow), I have no more
nerves to maintain this unittest. So I heavily simplified it to use only
a single thread in addition to the main thread. This is actually the
main use-case for SQLiteDatabase anyway, so we don't loose a lot of
important test coverage.