Board editor: Do not reject placing vias on top of other items
In the "add via" tool, do not reject placing the via anymore if there
are pads or vias at the cursor position, or traces of other net signals.
It could be annoying usability if the tool rejects to place a via in
some situations. For example placing a via in a high density area might
not be possible without creating short circuits. But often you still
want to place the via (temporarily creating a short circuit) and move
traces away afterwards. The tool should not reject you from doing this.
If you do it and don't realize that you created a short circuit, the
DRC will raise an error anyway when running the next time.