FabricationOutputDialog: Always use *.drl suffix for drill files
So far we used the *.drl file extension for drill files when choosing
the default settings, but *.txt when using Protel settings. However, it
seems that other tools use the *.drl file extension for drills even
when using Protel file extensions. And Ucamco highly recommends to use
either *.nc or *.drl. Since *.drl seems to be more commonly used, let's
use it too when choosing the Protel settings in the fabrication output
dialog.