ExcellonGenerator: Set Gerber file/apterure attributes
Although Excellon files are not Gerber files, there is a standard [1]
describing how Gerber X2 attributes can also be used in Excellon files
to add metadata in a standardized way. This might help some fabricators
to automatically detect whether drills are plated or not, and between
which layers the drills need to be created.

[1] https://www.ucamco.com/files/downloads/file_en/305/the-xnc-file-format-specification-technical-manual_en.pdf