Show error message if invalid workspace path specified
If the workspace path to open (e.g. loaded from user settings) does not
point to a valid workspace, an error message will be shown now.
Previously, the "first run"-wizard has been shown silently in this case,
so the user didn't notice that there was something wrong with the
specified workspace path.