Ask for new workspace path if workspace failed to open
Instead of closing the application with an error message if the
workspace failed to open, now the "first run"-wizard will be shown
after the error message box so the user is able to choose another
workspace directory.

Previously the user had to manually reset the workspace path stored
in the user settings ($HOME/.config/LibrePCB/) if it pointed to an
existing, but somehow broken workspace (e.g. not writable).