CI: Print various information about environment
It is very common to temporarily add some print statements to the CI
scripts when debugging CI problems. But since they are often a bit
hacky and useful only for a very specific problem, they are removed
once the problem is fixed.

To avoid adding/removing such print statements in future, I have now
added a new build step which prints a lot of information about the CI
environment (PATH, Qt version, Python version, OpenSSL version, ...)
and which is clean and generic enough to add it permanently to the
repository ;-)