CMake: Add WIN32 flag to GUI applications to hide cmd.exe
When building on Windows, our executables need to be marked as GUI
applications, otherwise the cmd.exe window will be visible beside
the LibrePCB windows. With qmake this was the default (opt-out with
"CONFIG += console") but now with CMake we have to explicitly configure
this (opt-in with WIN32 flag).

See https://cmake.org/cmake/help/latest/command/add_executable.html