Board editor: Fix unintended layer change when drawing trace
When drawing traces, the layer sometimes switched to another layer when
clicking on a position where another trace was already present. This
happened because the last drawn trace segment was not memorized before
deciding where to start the next segment. Now it is memorized so the
next segment should always start where the last segment ended.