Board editor: Fix layer change when starting at via or pad
When starting to draw a new trace segment at a via or THT pad, sometimes
it was not possible to change the layer, because the trace segment was
deleted and a new one created, but its start point could have moved to
another, very close anchor instead of the original starting via or pad.
Now the original starting via or pad is used when switching the layer.