CMake: Allow in-source builds
Although for development we usually work with out-of-source builds to
not cluttering up the whole source directory with output files, there
are situations where in-source builds are totally fine, or even much
easier to use than out-of-source builds.

For example, Snapcraft uses in-source builds and it would be a nightmare
to manually override the build commands to use out-of-source builds.
Maybe there are also other packaging systems which have some kind of
pre-defined CMake build workflow where out-of-source builds are not
supported. You can also imagine building LibrePCB in a Docker container
where it is just irrelevant if the output files are created in-source.

I think the risk of annoying developers by accidentally doing an
in-source build is not as high as the annoyance of package maintainers
and users who only want to compile it (no development). Especially
since the number of developers is quite small ;-) In addition, our
official build instructions only show commands for out-of-source builds,
so for those following the instructions, it doesn't matter if in-source
builds are possible or not.