Board editor: Fix modifying rotation of mirrored devices
When manually entering a new rotation value for a device in the device
properties dialog, the a wrong rotation was applied for mirrored
(bottom side) devices. Only for non-mirrored (top side) devices the
rotation worked properly.

Fixed by considering the inverted rotation behavior for mirrored
devices in the corresponding undo command.