AddComponentDialog: Fix placement of "Add more" checkbox
The "Add more" checkbox placement position was managed by
QDialogButtonBox and was therefore depending on the operating system.
On Windows, it was placed between the "Ok" and "Cancel" buttons, which
is very strange.

Now managing the layout manually with QLayout to ensure the checkbox
is always on the left side of both buttons.