CMake: Add alias for muparser target
Building LibrePCB with unbundled MuParser on Arch Linux currently fails,
because the target `MuParser::MuParser` is not defined. Because this
used to work before, I assume the target was renamed from
`MuParser::MuParser` to `muparser::muparser` some time ago.

To fix this, add an alias if the CMake package was found, but the pascal
case alias isn't defined yet.