Implement attributes "BOARD_INDEX" and "BOARD_DIRNAME"
These new built-in attributes of boards are especially useful as
placeholders in output filepaths passed to the CLI. The currently
often used placeholder `{{BOARD}}` is not absolutely guaranteed to be
unique across all boards because board names might contain characters
not allowed in filepaths, which are then removed or replaced. So
theoretically it could lead to conflicting output files when
exporting multiple boards (e.g. Gerber export).

With the new attributes, the placeholders `{{BOARD_INDEX}}` and
`{{BOARD_DIRNAME}}` are guaranteed to be unique and do not contain
illegal filepath characters, so they allow to export boards without
the risk of conflicts.