Automatically abort blocking tool in other editor
When you start a tool in either the schematic- or board editor which
blocks the undo command stack (due to intermediate changes, e.g. while
drawing a line), and then starting such a tool in the other editor
(e.g. switching from schematic- to board editor), currently you get
the error "Another command is active at the moment." so you had to
exit the other tool first.

This can be a bit annoying and confusing. It would be much more
convenient if the blocking tool gets aborted automatically and no
error appears. So this commit applies this change.