Fix modifying selection when clicking fast
When clicking fast while holding Shift to cycle the selection or Ctrl
to invert the selection, the event was handled like a double click and
thus opened the properties dialog instead of modifying the selection.
Fixed in all editors by handling double click identical to single click
if Shift or Ctrl is pressed.