Keep selection when clicking on selected item
Currently the selection was reset when clicking without Shift or Ctrl
on an item, for example when double-clicking to open the properties
dialog. Therefore it was not possible to open the properties dialog
on an item other than the top-most item.

Fixed by keeping the selection when clicking on an already selected
item, no matter whether it's the top-most item or not. This allows to
select an item with the Shift modifier, and then double-clicking on
the selected item.