Scale cursor click tolerance with zoom level
Instead of using a fixed physical tolerance when finding items below
the cursor when clicking, use a dynamic, zoom-dependent tolerance which
leads to a constant tolerance in terms of screen pixels. This way all
items can be grabbed easily (even zero-width polygons), independent of
the zoom level.

In addition, the algorithm to determine the clicked item now prioritizes
items *exactly* below the cursor over items which are only *close* to
the cursor. And for those items, even the distance to the cursor it
taken into account. This leads to a more intuitive selection order when
cycling through the items with Shift, and often even eliminates the
need for pressing Shift since the desired item is selected just with
the first click.