Show waiting spinner in tree/list views during library scan
In all tree views and list views which show content from the workspace
library database (e.g. categories, symbols, components, ...), show a
spinner during the scan to indicate that the content might be outdated
and will be updated soon.

This should help to avoid confusion when users add libraries or library
elements, but then can't find those things in the tree/list views.
Especially after installing a new library, the whole library seems to
be empty for quite some time - with the spinner, it should be clear
that some update is in progress.