Immediately show library scan progress bar in windows
If the workspace library background scan is in progress while a new
window is opened, the progress bar in the status bar did not appear
immediately but only on the next progress update event (could be several
seconds later).

Now the progress bar is shown immediately when opening a window while
the scan is in progress.