Schematic editor: Show info banner if schematic is empty
When opened an empty schematic (which is the case after creating a new
project), show a banner in the schematic editor with a hint how to add
components to the schematic. The banner sugests to first add a schematic
frame and provides a hyperlink to directly open the "Add Component"
dialog with components filtered by the term "schematic frame". If the
official base library is installed, the user will directly see the
available schematic frames.

Intended to help users getting started with LibrePCB.