Fix possibly unexpected center of rotate/mirror operations
When rotating/mirroring items in any of the editors, so far the center
point of the operation was always mapped to the grid interval. However,
this is not expected when only a single item is selected since it moves
the item's position if it was not aligned to the grid. Thus now using
the item's exact center position (not mapped to grid) if only one is
selected. For example rotating a single, unaligned pad now keeps its
exact position.

If more than one item is selected, the behavior is not affected by this
change.

This also seems to fix a bug (#728) of possibly unaligning a previously
aligned device.