Gerber export: Enable solder paste export by default
If you forget to enable the top/bottom solder paste export in the
Gerber export settings but still want to order a stencil, there's a
risk that the PCB manufacturer auto-generates the stencil from the
stop mask layer (or from the copper layer?). But the other way around,
if you don't need a stencil but forget to disable exporting these
layers, I think there's no risk at all (just two unnecessary files
generated). So changing from opt-in to opt-out is probably the safer
option.