Move board related messages from ERC to DRC
Although the ERC is nicer than the DRC because it is instantly updated,
board related messages (e.g. unplaced component) are usually not
of interest in the schematic editor. And in the board editor, probably
nobody checks the ERC messages. So let's move these messages to the DRC
to strictly distinguish between board issues (-> DRC) and
schematic/circuit issues (-> ERC).