Rename board device property "mirror" to "flip"
For consistent terminology with the "Flip" action (Ctrl+F) in the
board editor, rename the property "mirror" of devices to "flip". This
is more intuitive because flip is not the same as mirror (in contrast
to schematics, where mirror is indeed just a graphical operation).

However, in the source code it's still called "mirror" for
compatibility with the `Transform` template constructor.