Board editor: Render pad & via texts with Fontobene
So far the texts within pads and vias (pad number resp. net name) were
drawn with a TrueType font, but it turned out this rendering was a
huge performance bottleneck, slowing down the whole board editor on
large designs. Now rendering them with a Fontobene stroke font which
is much faster.