WorkspaceLibraryScanner: Do not warn about empty directories
So far, the background library scanner emitted warnings to `stderr` for
each directory within workspace libraries not containing a valid
library element. This is useful to locate issues if a library element
does not appear in the library editor for some reason.

However, especially when working with Git it happens often that empty
directories are left over, leading to such warnings although there's
nothing bad going on. Thus now only emitting these warnings on
*non-empty* directories which do not contain valid library elements.