Suppress warning on stderr caused by argv[0]
LibrePCB tries to open all project files passed as arguments to the
process. If the passed files do not have the extension `*.lpp` or
`*.lppz`, a warning is printed on `stderr` (useful for debugging
issues if opening projects does not work).

However, `argv[0]` usually (but not absolutely always) contains the
path to the application executable, causing such a warning to be
printed which is a bit ugly. Thus now suppressing the warning on
`argv[0]`, but still printing it on any additional argument.