Fix possibly broken copy&paste in editors on some systems
At least on one system the clipboard didn't work if no text was
set, so let's also copy the `SExpression` as `text/plain` (beside the
binary data) into the clipboard as a workaround. Might even be just a
temporary bug in the clipboard manager of my system but it's hard to
find the root cause and being able to paste the `SExpression` into
a text editor might be useful anyway, e.g. for debugging purposes.