Show waiting spinner while opening external applications
After triggering an external application from within LibrePCB (web
browser, PDF viewer, ...), change the cursor to a waiting spinner for
a moment to give immediate feedback that the operation has been
successfully triggered. This is a much better user experience than
silently waiting seconds for the requested application to appear.