BOM export dialog: Don't scale columns with window width
Currently the width of most columns in the BOM export dialog was
scaled with the window width to avoid a horizontal scrollbar to
appear. However, usually this caused cropped text in columns with
long texts and unused space in columns with short texts. Thus now
changed the behavior to always fit all columns to their content and
showing a horizontal scrollbar if not all columns fit into the
window. In addition, the columns could manually be resized when
needed (not memorized though since the optimal width depends on the
content).