Schematic editor: Hide redundant pin numbers
Hide pin number if it's identical to the pin name shown on the
schematic. This avoids cluttering the schematic with redundant
information, mainly for components like connectors where pin names
and pad names are just numbers (1, 2, 3, ...). In such cases, hiding
pin numbers also reduces the risk of overlaps in case of non-standard
pin lengths (also often the case for connectors).