Schematic editor: Improve selection of stacked items
If text labels are located on top of a symbol, make a click selecting
the text instead of the symbol (consistent order of drawing and
selection). With one exception: If the symbols origin is below the
cursor, the symbol is selected instead of the text. This is required
to allow selecting zero-area symbols like GND or VCC.