Give GUI higher thread priority than background tasks
In some experiments with QtQuick I observed significant lags in the GUI
when the background library scanner was running, which seem to be fixed
by giving the main thread (GUI) a higher priority than the library
scanner. Though not sure if relevant with QtWidgets (not using QtQuick
for now) and some systems might not support thread priorities, I think
it makes sense anyway to assign thread priorities. Thus also assigned
the network thread a lower priority than the main thread too.