Library editor: Fix conflicting component signal names
When creating a new component in the library editor which consists of
multiple gates, the automatically added component signals may have
conflicting names (e.g. 2x "OUT" for a dual opamp). This change appends
an incrementing number in this case (e.g. "OUT1" & "OUT2").

The "Automatically assign all signals by name" also takes this suffix
into account to get a successful auto-assignment, no matter if the
suffix is there or not.